+++ all.go
@@ -25,7 +25,8 @@ func All[T any](rules ...Rule[T]) Rule[T] { return &allRule[T]{children: rules} 
 
 type allRule[T any] struct{ children []Rule[T] }
 
-func (*allRule[T]) rule() {}
+func (*allRule[T]) rule()                      {}
+func (*allRule[T]) inspectionStrategy() string { return "all" }
 func (r *allRule[T]) newState(ids *nodeIDAllocator, hints *buildStatistics) Rule[T] {
 	children := make([]Rule[T], len(r.children))
 	for i, child := range r.children {